requirements.txt
openai
e2b
e2b-code-interpreter
python-dotenv